Interdire la suppression des configs utilisées

Closes #2302

See merge request !66

+++ server/web2py/applications/vidjil/controllers/config.py
@@ -16,9 +16,12 @@ def index():
     
 
     query = db((auth.vidjil_accessible_query(PermissionEnum.read_config.value, db.config) | auth.vidjil_accessible_query(PermissionEnum.admin_config.value, db.config) ) ).select(orderby=~db.config.name)
+    used_query = db(db.results_file.config_id > 0).select(db.results_file.config_id, distinct=True)
+    used_configs = [row.config_id for row in used_query]
 
     return dict(message=T('Configs'),
                query=query,
+               used_configs=used_configs,
                isAdmin = auth.is_admin())
 
 
@@ -115,15 +118,19 @@ def confirm():
 
 def delete():
     if (auth.can_modify_config(request.vars['id'])):
-        #delete results_file using this config
-        db(db.results_file.config_id==request.vars["id"]).delete()
+        count = db(db.results_file.config_id==request.vars["id"]).count()
 
-        #delete config
-        db(db.config.id==request.vars["id"]).delete()
+        if (count == 0):
+            #delete config
+            db(db.config.id==request.vars["id"]).delete()
 
-        res = {"redirect": "config/index",
-               "message": "config '%s' deleted" % request.vars["id"]}
-        log.admin(res)
+            res = {"redirect": "config/index",
+                   "message": "config '%s' deleted" % request.vars["id"]}
+            log.admin(res)
+        else:
+            res = {"redirect": "config/index",
+                    "success": "false",
+                    "message": "cannot delete a config that has been used"}
         return gluon.contrib.simplejson.dumps(res, separators=(',',':'))
     return error_message(ACCESS_DENIED)
